Deepfake Detector/requirements.txt

torch
opencv-python
numpy
scikit-learn
Pillow
imageio
ffmpeg-python
matplotlib
joblib
flask
transformers
torchvision
pillow